inputs func components
